requirements.txt
# IBM watsonx Orchestrate ADK
ibm-watsonx-orchestrate>=1.0.0

# External API clients
notion-client>=2.2.1
pyairtable>=2.3.3
slack-sdk>=3.27.1

# Data handling and validation
pydantic>=2.6.0
python-dotenv>=1.0.1

# HTTP requests
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.2
